Created new TabAnimation class
