+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/AnimatedTab.kt
@@ -5,7 +5,7 @@ import javafx.scene.Node
 import javafx.scene.control.Tab
 import tornadofx.*
 
-abstract class ChromeableTab : Tab() {
+abstract class AnimatedTab : Tab() {
     val animatedContentProperty = SimpleObjectProperty<Node>()
     var animatedContent: Node by animatedContentProperty
