+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/TabAnimation.kt
@@ -1,0 +1,78 @@
+package org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage
+
+import com.jfoenix.transitions.CachedTransition
+import javafx.animation.Interpolator
+import javafx.animation.KeyFrame
+import javafx.animation.KeyValue
+import javafx.animation.Timeline
+import javafx.scene.Node
+import javafx.scene.control.TabPane
+import javafx.util.Duration
+
+class TabAnimation(private val tabPane: TabPane) {
+
+    enum class TransitionDirection {
+        LEFT,
+        RIGHT
+    }
+
+    init {
+        setTabSelectionAnimation()
+    }
+
+    private fun setTabSelectionAnimation() {
+        tabPane.selectionModel.selectedIndexProperty().addListener { _, old, new ->
+            val oldIndex = old.toInt()
+            val newIndex = new.toInt()
+            if (oldIndex >= 0 && newIndex >= 0) {
+                val direction = if (oldIndex > newIndex) TransitionDirection.RIGHT else TransitionDirection.LEFT
+                val tab: AnimatedTab? = tabPane.tabs[newIndex] as? AnimatedTab
+                if (tab != null) {
+                    animate(tab.animatedContent, direction)
+                }
+            }
+        }
+    }
+
+    // Animate first tab's content
+    fun animate(direction: TransitionDirection) {
+        if (tabPane.tabs.size > 0) {
+            if (tabPane.selectionModel.selectedIndex > 0) {
+                tabPane.selectionModel.select(0)
+            } else {
+                val tab: AnimatedTab? = tabPane.tabs[0] as? AnimatedTab
+                if (tab != null) {
+                    animate(tab.animatedContent, direction)
+                }
+            }
+        }
+    }
+
+    private fun animate(content: Node, direction: TransitionDirection) {
+        val contentWidth = when (direction) {
+            TransitionDirection.LEFT -> tabPane.width
+            TransitionDirection.RIGHT -> -tabPane.width
+        }
+
+        content.translateX = contentWidth
+
+        object : CachedTransition(
+            content,
+            Timeline(
+                KeyFrame(
+                    Duration.millis(1000.0),
+                    KeyValue(
+                        content.translateXProperty(),
+                        0.0,
+                        Interpolator.EASE_BOTH
+                    )
+                )
+            )
+        ) {
+            init {
+                cycleDuration = Duration.seconds(0.320)
+                delay = Duration.seconds(0.0)
+            }
+        }.play()
+    }
+}
