+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/tabgroups/ChapterTabGroup.kt
@@ -3,7 +3,7 @@ package org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.tabgroup
 import javafx.scene.control.Tab
 import org.wycliffeassociates.otter.common.data.model.ResourceMetadata
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.cardgrid.view.CardGridFragment
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.ChromeableTab
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.AnimatedTab
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.workbook.viewmodel.WorkbookViewModel
 import tornadofx.*
 
@@ -52,7 +52,7 @@ class ChapterTabGroup : TabGroup() {
         tabMap[resourceMetadata.identifier]?.select()
     }
 
-    private inner class ChapterSelectTab(val resourceMetadata: ResourceMetadata) : ChromeableTab() {
+    private inner class ChapterSelectTab(val resourceMetadata: ResourceMetadata) : AnimatedTab() {
         init {
             text = resourceMetadata.identifier
             add(CardGridFragment().root)
