+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/chromeablestage/tabgroups/RecordScriptureTabGroup.kt
@@ -1,13 +1,13 @@
 package org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.tabgroups
 
-import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.ChromeableTab
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.chromeablestage.AnimatedTab
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.takemanagement.view.RecordScriptureFragment
 import tornadofx.*
 
 class RecordScriptureTabGroup : TabGroup() {
     private val tab = RecordScriptureTab()
 
-    private inner class RecordScriptureTab : ChromeableTab() {
+    private inner class RecordScriptureTab : AnimatedTab() {
         init {
             add(RecordScriptureFragment().root)
         }
