+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/resourcetakes/view/RecordResourceFragment.kt
@@ -132,6 +132,7 @@ class RecordResourceFragment(
     }
 
     private val navbar = hbox {
+        id = "navbar"
         addClass(RecordResourceStyles.navbar)
         anchorpaneConstraints {
             bottomAnchor = 0.0
